requirements.txt
streamlit==1.41.1
python-gnupg==0.5.3
paramiko==3.5.0
cryptography==44.0.0
pyopenssl==24.3.0
pytest==8.3.4
pytest-cov==6.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
pydantic-extra-types>=2.6.0
email-validator>=2.1.0
python-multipart>=0.0.5
